rdic_pipeline/requirements-mac.txt
numpy<2
pandas>=2.1,<3
scipy>=1.10,<2
scikit-learn>=1.3,<2
tqdm
pyarrow
networkx
matplotlib

transformers>=4.39,<5
datasets>=2.18,<3
evaluate
accelerate
tokenizers
safetensors

gymnasium>=0.29,<1
stable-baselines3>=2.3,<3
tensorboard
joblib
pyyaml
